compiler warning suppression

+++ clamd/clamd.c
@@ -361,7 +361,7 @@ int main(int argc, char **argv)
 
 
         if(logg_size)
-            logg("#Log file size limited to %u bytes.\n", logg_size);
+            logg("#Log file size limited to %lld bytes.\n", (long long int)logg_size);
         else
             logg("#Log file size limit disabled.\n");
