+++ libclamav/tomsfastmath/misc/fp_ident.c
@@ -18,8 +18,8 @@ const char *fp_ident(void)
 "TomsFastMath (%s)\n"
 "\n"
 "Sizeofs\n"
-"\tfp_digit = %lu\n"
-"\tfp_word  = %lu\n"
+"\tfp_digit = %llu\n"
+"\tfp_word  = %llu\n"
 "\n"
 "FP_MAX_SIZE = %u\n"
 "\n"
@@ -70,7 +70,7 @@ const char *fp_ident(void)
 #ifdef TFM_HUGE
 " TFM_HUGE "
 #endif
-"\n", __DATE__, sizeof(fp_digit), sizeof(fp_word), FP_MAX_SIZE);
+            "\n", __DATE__, (long long unsigned)sizeof(fp_digit), (long long unsigned)sizeof(fp_word), FP_MAX_SIZE);
 
    if (sizeof(fp_digit) == sizeof(fp_word)) {
       strncat(buf, "WARNING: sizeof(fp_digit) == sizeof(fp_word), this build is likely to not work properly.\n",
